tor: Refactor into separate files for forms and views.
