+++ plinth/modules/tor/__init__.py
@@ -16,19 +16,176 @@
 #
 
 """
-Plinth module to configure Tor
+Plinth module to configure Tor.
 """
 
+import augeas
 from django.utils.translation import ugettext as _
+import glob
+import itertools
 
-from . import tor
-from .tor import init
 from plinth import actions
 from plinth import action_utils
+from plinth import cfg
+from plinth.modules.names import SERVICES
+from plinth.signals import domain_added
 
-__all__ = ['tor', 'init']
 
 depends = ['plinth.modules.apps', 'plinth.modules.names']
+
+APT_SOURCES_URI_PATHS = ('/files/etc/apt/sources.list/*/uri',
+                         '/files/etc/apt/sources.list.d/*/*/uri')
+APT_TOR_PREFIX = 'tor+'
+
+
+def init():
+    """Initialize the module."""
+    menu = cfg.main_menu.get('apps:index')
+    menu.add_urlname(_('Anonymity Network (Tor)'), 'glyphicon-eye-close',
+                     'tor:index', 100)
+
+    # Register hidden service name with Name Services module.
+    (hs_enabled, hs_hostname, hs_ports) = get_hs()
+
+    if is_enabled() and is_running() and hs_enabled and hs_hostname:
+        hs_services = []
+        for service in SERVICES:
+            if str(service[2]) in hs_ports:
+                hs_services.append(service[0])
+    else:
+        hs_hostname = None
+        hs_services = None
+
+    domain_added.send_robust(
+        sender='tor', domain_type='hiddenservice',
+        name=hs_hostname, description=_('Tor Hidden Service'),
+        services=hs_services)
+
+
+def is_enabled():
+    """Return whether the module is enabled."""
+    return action_utils.service_is_enabled('tor')
+
+
+def is_running():
+    """Return whether the service is running."""
+    return action_utils.service_is_running('tor')
+
+
+def get_status():
+    """Return current Tor status."""
+    output = actions.superuser_run('tor', ['get-ports'])
+    port_info = output.split('\n')
+    ports = {}
+    for line in port_info:
+        try:
+            (key, val) = line.split()
+            ports[key] = val
+        except ValueError:
+            continue
+
+    (hs_enabled, hs_hostname, hs_ports) = get_hs()
+
+    return {'enabled': is_enabled(),
+            'is_running': is_running(),
+            'ports': ports,
+            'hs_enabled': hs_enabled,
+            'hs_hostname': hs_hostname,
+            'hs_ports': hs_ports,
+            'apt_transport_tor_enabled': is_apt_transport_tor_enabled()}
+
+
+def get_hs():
+    """Return hidden service status."""
+    output = actions.superuser_run('tor', ['get-hs'])
+    output = output.strip()
+    if output == '':
+        hs_enabled = False
+        hs_hostname = 'Not Configured'
+        hs_ports = ''
+    elif output == 'error':
+        hs_enabled = False
+        hs_hostname = 'Not available (Run Tor at least once)'
+        hs_ports = ''
+    else:
+        hs_enabled = True
+        hs_info = output.split()
+        hs_hostname = hs_info[0]
+        hs_ports = hs_info[1]
+
+    return (hs_enabled, hs_hostname, hs_ports)
+
+
+def get_augeas():
+    """Return an instance of Augeaus for processing APT configuration."""
+    aug = augeas.Augeas(flags=augeas.Augeas.NO_LOAD +
+                        augeas.Augeas.NO_MODL_AUTOLOAD)
+    aug.set('/augeas/load/Aptsources/lens', 'Aptsources.lns')
+    aug.set('/augeas/load/Aptsources/incl[last() + 1]', '/etc/apt/sources.list')
+    aug.set('/augeas/load/Aptsources/incl[last() + 1]',
+            '/etc/apt/sources.list.d/*.list')
+    aug.load()
+
+    # Currently, augeas does not handle Deb822 format, it error out.
+    if aug.match('/augeas/files/etc/apt/sources.list/error') or \
+       aug.match('/augeas/files/etc/apt/sources.list.d//error'):
+        raise Exception('Error parsing sources list')
+
+    # Starting with Apt 1.1, /etc/apt/sources.list.d/*.sources will
+    # contain files with Deb822 format.  If they are found, error out
+    # for now.  XXX: Provide proper support Deb822 format with a new
+    # Augeas lens.
+    if glob.glob('/etc/apt/sources.list.d/*.sources'):
+        raise Exception('Can not handle Deb822 source files')
+
+    return aug
+
+
+def iter_apt_uris(aug):
+    """Iterate over all the APT source URIs."""
+    return itertools.chain.from_iterable([aug.match(path)
+                                          for path in APT_SOURCES_URI_PATHS])
+
+
+def get_real_apt_uri_path(aug, path):
+    """Return the actual path which contains APT URL.
+
+    XXX: This is a workaround for Augeas bug parsing Apt source files
+    with '[options]'.  Remove this workaround after Augeas lens is
+    fixed.
+    """
+    uri = aug.get(path)
+    if uri[0] == '[':
+        parent_path = path.rsplit('/', maxsplit=1)[0]
+        skipped = False
+        for child_path in aug.match(parent_path + '/*')[1:]:
+            if skipped:
+                return child_path
+
+            value = aug.get(child_path)
+            if value[-1] == ']':
+                skipped = True
+
+    return path
+
+
+def is_apt_transport_tor_enabled():
+    """Return whether APT is set to download packages over Tor."""
+    try:
+        aug = get_augeas()
+    except Exception:
+        # If there was an error with parsing or there are Deb822
+        # files.
+        return False
+
+    for uri_path in iter_apt_uris(aug):
+        uri_path = get_real_apt_uri_path(aug, uri_path)
+        uri = aug.get(uri_path)
+        if not uri.startswith(APT_TOR_PREFIX) and \
+           (uri.startswith('http://') or uri.startswith('https://')):
+            return False
+
+    return True
 
 
 def diagnose():
