+++ plinth/modules/tor/views.py
@@ -1,0 +1,127 @@
+#
+# This file is part of Plinth.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+#
+
+"""
+Plinth module for configuring Tor.
+"""
+
+from django.contrib import messages
+from django.template.response import TemplateResponse
+from django.utils.translation import ugettext_lazy as _
+
+from .forms import TorForm
+from plinth import actions
+from plinth import action_utils
+from plinth import package
+from plinth.errors import ActionError
+from plinth.modules import tor
+from plinth.modules.names import SERVICES
+from plinth.signals import domain_added, domain_removed
+
+
+def on_install():
+    """Setup Tor configuration as soon as it is installed."""
+    actions.superuser_run('tor', ['setup'])
+    actions.superuser_run('tor', ['enable-apt-transport-tor'])
+
+
+@package.required(['tor', 'tor-geoipdb', 'torsocks', 'obfs4proxy',
+                   'apt-transport-tor'],
+                  on_install=on_install)
+def index(request):
+    """Serve configuration page."""
+    status = tor.get_status()
+
+    form = None
+
+    if request.method == 'POST':
+        form = TorForm(request.POST, prefix='tor')
+        # pylint: disable=E1101
+        if form.is_valid():
+            _apply_changes(request, status, form.cleaned_data)
+            status = tor.get_status()
+            form = TorForm(initial=status, prefix='tor')
+    else:
+        form = TorForm(initial=status, prefix='tor')
+
+    return TemplateResponse(request, 'tor.html',
+                            {'title': _('Tor Control Panel'),
+                             'status': status,
+                             'form': form})
+
+
+def _apply_changes(request, old_status, new_status):
+    """Try to apply changes and handle errors."""
+    try:
+        __apply_changes(request, old_status, new_status)
+    except ActionError as exception:
+        messages.error(request, _('Action error: {0} [{1}] [{2}]').format(
+            exception.args[0], exception.args[1], exception.args[2]))
+
+
+def __apply_changes(request, old_status, new_status):
+    """Apply the changes."""
+    if old_status['enabled'] == new_status['enabled'] and \
+       old_status['hs_enabled'] == new_status['hs_enabled'] and \
+       old_status['apt_transport_tor_enabled'] == \
+       new_status['apt_transport_tor_enabled']:
+        messages.info(request, _('Setting unchanged'))
+        return
+
+    if old_status['enabled'] != new_status['enabled']:
+        if new_status['enabled']:
+            actions.superuser_run('tor', ['enable'])
+            messages.success(request, _('Tor enabled'))
+        else:
+            actions.superuser_run('tor', ['disable'])
+            messages.success(request, _('Tor disabled'))
+
+    if old_status['hs_enabled'] != new_status['hs_enabled']:
+        if new_status['hs_enabled']:
+            actions.superuser_run('tor', ['enable-hs'])
+            messages.success(request, _('Tor hidden service enabled'))
+        else:
+            actions.superuser_run('tor', ['disable-hs'])
+            messages.success(request, _('Tor hidden service disabled'))
+
+    # Update hidden service name registered with Name Services module.
+    domain_removed.send_robust(
+        sender='tor', domain_type='hiddenservice')
+
+    enabled = action_utils.service_is_enabled('tor')
+    is_running = action_utils.service_is_running('tor')
+    (hs_enabled, hs_hostname, hs_ports) = tor.get_hs()
+
+    if enabled and is_running and hs_enabled and hs_hostname:
+        hs_services = []
+        for service in SERVICES:
+            if str(service[2]) in hs_ports:
+                hs_services.append(service[0])
+
+        domain_added.send_robust(
+            sender='tor', domain_type='hiddenservice',
+            name=hs_hostname, description=_('Tor Hidden Service'),
+            services=hs_services)
+
+    if old_status['apt_transport_tor_enabled'] != \
+       new_status['apt_transport_tor_enabled']:
+        if new_status['apt_transport_tor_enabled']:
+            actions.superuser_run('tor', ['enable-apt-transport-tor'])
+            messages.success(request, _('Enabled package download over Tor'))
+        else:
+            actions.superuser_run('tor', ['disable-apt-transport-tor'])
+            messages.success(request, _('Disabled package download over Tor'))
